feat: add light sdk read and resolver surfaces
